Move Project login settings to its own tab (#407)
![Screenshot 2025-05-19 at 1 11
06 PM](https://github.com/user-attachments/assets/2d70af24-d220-43fd-ab2c-281a577f6c85)
![Screenshot 2025-05-19 at 1 12
06 PM](https://github.com/user-attachments/assets/2e6c9714-2db0-47b6-b98f-dd85c67170e1)

---------
